// hvn-telegram-bot/src/shared/types/data.ts
export type Activity = {
  id: string; // Firestore document ID
  srNo: number;
  employeeName: string;
  action: string;
  description: string;
  timestamp: Timestamp;
  createdBy: string;
  source: 'UI' | 'BOT';
  groupName?: string;
};